Move exec shell/debug into its own dir and subcommand
Create a new exec sub-app under the main `deployer` app and create two
more apps under it called `shell` and `debug`.
 - Move the old deployer/debug.py code under
   deployer/exec/debug/app_and_commands.py and register some of the functions there
   under the `deployer exec debug` command.
 - Move the deployer/cloud_access.py under
   deployer/exec/shell/cloud_commands.py and register the functions
   there under the `deployer exec shell` command.
 - Move the two functions related to shell execution inside the hub and
   homes dirs from the debug.py file into
   deployer/exec/shell/infra_components_commands.py
